backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.8.2
pydantic-settings==2.3.0
sqlalchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
clerk-backend-api==1.1.0
supabase==2.4.6
python-multipart==0.0.9
stripe==9.12.0
anthropic>=0.84.0
openpyxl==3.1.5
slowapi==0.1.9
resend==2.4.0
fpdf2==2.7.9
qrcode[pil]==7.4.2
